WIP testos i millorar funcionament

+++ som_crawlers/wizard/wizard_executar_tasca.py
@@ -1,6 +1,7 @@
 from ast import Param
 import base64
 from fileinput import filename
+from genericpath import isfile
 from quopri import encodestring
 from ssl import DefaultVerifyPaths
 import pooler
@@ -23,7 +24,6 @@ class WizardExecutarTasca(osv.osv_memory):
     def executar_tasca(self, cursor, uid, ids, context=None):
         #obtenim l'objecte tasca
         task = self.pool.get('som.crawlers.task')
-
 
         if not context:
             return False
@@ -53,38 +53,44 @@ class WizardExecutarTasca(osv.osv_memory):
         taskStepParams = json.loads(taskStep_obj.params)
         path = os.path.dirname(os.path.realpath(__file__))
 
+        result_id = classresult.create(cursor,uid,{'task_id': taskStep_obj.task_id.id, 'name': taskStep_obj.name})
+
         if taskStepParams.has_key('nom_fitxer'):
             config_obj=self.id_del_portal_config(cursor,uid,id,context)
             filePath = os.path.join(path, "../scripts/" + taskStepParams['nom_fitxer'])
-            path_python = "~/.virtualenvs/massive/bin/python"
-            fileName = "output_" + config_obj.name + "_" + datetime.now().strftime("%Y-%m-%d_%H_%M") + ".txt"
-            os.system(path_python + " " + filePath + " -n "+ config_obj.name + " -u " + config_obj.usuari + " -p " + config_obj.contrasenya + " -f " + fileName + " -url " + config_obj.url_portal + " -fltr " + config_obj.filtres + " -c " + config_obj.crawler + " -d " + str(config_obj.days_of_margin) + " -nfp " + str(config_obj.pending_files_only) + " -b " + config_obj.browser) 
-            result_id = classresult.create(cursor,uid,{'task_id': taskStep_obj.task_id.id})
-            with open(os.path.join(path,"../outputFiles",fileName)) as f:
-                output = f.read().replace('\n', ' ')
+            if os.path.exists(filePath):
+                path_python = "~/.virtualenvs/massive/bin/python"
+                fileName = "output_" + config_obj.name + "_" + datetime.now().strftime("%Y-%m-%d_%H_%M") + ".txt"
+                os.system(path_python + " " + filePath + " -n "+ config_obj.name + " -u " + config_obj.usuari + " -p " + config_obj.contrasenya + " -f " + fileName + " -url " + config_obj.url_portal + " -fltr " + config_obj.filtres + " -c " + config_obj.crawler + " -d " + str(config_obj.days_of_margin) + " -nfp " + str(config_obj.pending_files_only) + " -b " + config_obj.browser) 
 
-            os.remove(os.path.join(path, "../outputFiles",fileName))
-            if output == 'Files have been successfully downloaded':
+                with open(os.path.join(path,"../outputFiles",fileName)) as f:
+                    output = f.read().replace('\n', ' ')
 
-               self.attach_files_zip(cursor, uid, result_id, config_obj, path, context = context)
+                os.remove(os.path.join(path, "../outputFiles",fileName))
+                if output == 'Files have been successfully downloaded':
+                    output = self.attach_files_zip(cursor, uid, id, result_id, config_obj, path, context = context)
 
+                f.close()
+            else:
+                output = 'File or directory doesn\'t exist'
         else:
             output = 'Falta especificar nom fitxer'
 
         data_i_hora = datetime.now().strftime("%Y-%m-%d_%H:%M")
         taskStep_obj.task_id.write({'ultima_tasca_executada': str(taskStep_obj.task_id.name)+ ' - ' + str(data_i_hora)})
         classresult.write(cursor,uid, result_id, {'data_i_hora_execucio': data_i_hora, 'resultat': output})
-        f.close()
 
+        return output
 
     def import_xml_files(self, cursor, uid,id,context=None):
 
         classresult = self.pool.get('som.crawlers.result')
         classTaskStep = self.pool.get('som.crawlers.task.step')
         taskStep_obj=classTaskStep.browse(cursor,uid,id)
-
+        taskStepParams = json.loads(taskStep_obj.params)
+        result_id = classresult.create(cursor,uid,{'task_id': taskStep_obj.task_id.id, 'name': taskStep_obj.name})
         attachment_obj = self.pool.get('ir.attachment')
-        attachment_id = attachment_obj.search(cursor, uid, [('res_model',"=",'som.crawlers.result'),('res_id', "=", taskStep_obj.task_id.id)])
+        attachment_id = attachment_obj.search(cursor, uid, [('res_model',"=",'som.crawlers.task'),('res_id', "=", taskStep_obj.task_id.id)])
         if not attachment_id:
             output = "don't exist id attachment"
 
@@ -97,10 +103,9 @@ class WizardExecutarTasca(osv.osv_memory):
 
         data_i_hora = datetime.now().strftime("%Y-%m-%d_%H:%M")
         taskStep_obj.task_id.write({'ultima_tasca_executada': str(taskStep_obj.task_id.name)+ ' - ' + str(data_i_hora)})
-        result_id = classresult.search(cursor, uid, [('task_id',"=",taskStep_obj.task_id.id),('data_i_hora_execucio', "=", data_i_hora)])
-        classresult.write(cursor,uid, result_id, {'resultat': output})
-
-
+        #result_id = classresult.search(cursor, uid, [('task_id',"=",taskStep_obj.task_id.id),('data_i_hora_execucio', "=", data_i_hora)])
+        classresult.write(cursor,uid, result_id, {'data_i_hora_execucio': data_i_hora, 'resultat': output})
+        return output
 
     def id_del_portal_config(self,cursor,uid,id,context=None):
 
@@ -114,29 +119,40 @@ class WizardExecutarTasca(osv.osv_memory):
         return config_obj
 
 
-    def attach_files_zip(self, cursor, uid, id, config_obj, path, context=None):
+    def attach_files_zip(self, cursor, uid, id, result_id, config_obj, path, context=None):
 
+        classTaskStep = self.pool.get('som.crawlers.task.step')
+        taskStep_obj=classTaskStep.browse(cursor,uid,id)
         classresult = self.pool.get('som.crawlers.result')
         path_to_zip = os.path.join(path,'../tmp',config_obj.name)
-        for fileName in os.listdir(path_to_zip):
-            with open(os.path.join(path_to_zip,fileName), 'r') as f:
-                content  = f.read()
-            full_path = os.path.join(path_to_zip,fileName)
-            os.remove(full_path)
+        if not os.path.exists(path_to_zip):
+            output = "zip directory doesn\'t exist"
+        else:
+            if len(os.listdir(path_to_zip)) == 0:
+                 output = "Directori doesn\'t contain any ZIP"
+            else:
+                for fileName in os.listdir(path_to_zip):
+                    with open(os.path.join(path_to_zip,fileName), 'r') as f:
+                        content  = f.read()
+                    full_path = os.path.join(path_to_zip,fileName)
+                    os.remove(full_path)
 
-            pool = pooler.get_pool(cursor.dbname)
+                    pool = pooler.get_pool(cursor.dbname)
 
-            attachment = {
-                'name':  fileName,
-                'datas': base64.encodestring(content),
-                'datas_fname': fileName,
-                'res_model': 'som.crawlers.result',
-                'res_id': id,
-            }
+                    attachment = {
+                        'name':  fileName,
+                        'datas': base64.encodestring(content),
+                        'datas_fname': fileName,
+                        'res_model': 'som.crawlers.task',
+                        'res_id': taskStep_obj.task_id.id,
+                    }
 
-            attachment_id = self.pool.get('ir.attachment').create(cursor, uid, attachment, context=context)
-            classresult.write(cursor,uid, id, {'zip_name': attachment_id})
-            cursor.commit()
+                    attachment_id = self.pool.get('ir.attachment').create(cursor, uid, attachment, context=context)
+                    classresult.write(cursor,uid, result_id, {'zip_name': attachment_id})
+                    cursor.commit()
+                    output = "files succesfully attached"
+
+        return output
 
     def attach_files_xml(self, cursor, uid, id, config_obj, path, context=None):
